Report and user management updated

+++ app/Http/Controllers/RolesController.php
@@ -1,10 +1,13 @@
 <?php
+
 namespace App\Http\Controllers;
+
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Spatie\Permission\Models\Role;
 use Spatie\Permission\Models\Permission;
 use Illuminate\Support\Facades\DB;
+
 class RolesController extends Controller
 {
     public function index()
@@ -23,15 +26,16 @@ class RolesController extends Controller
         $role = Role::create(['name' => $request->input('name')]);
         return redirect()->route('roles.index')->with('success', 'Role created successfully.');
     }
-    public function fetchrole($id){
+    public function fetchrole($id)
+    {
         $permissions = DB::table('role_has_permissions')
-        ->select('permissions.name')
-        ->join('permissions', 'role_has_permissions.permission_id', '=', 'permissions.id')
-        ->where('role_has_permissions.role_id', $id)
-        ->get();
+            ->select('permissions.name')
+            ->join('permissions', 'role_has_permissions.permission_id', '=', 'permissions.id')
+            ->where('role_has_permissions.role_id', $id)
+            ->get();
         // dd($permissions);
-        return response()->json(['permissions' => $permissions]);   
-}
+        return response()->json(['permissions' => $permissions]);
+    }
 
     public function edit(Role $role)
     {
@@ -53,24 +57,27 @@ class RolesController extends Controller
 
     public function updatePermissions(Request $request, Role $role)
     {
-        $permissions=$request->permissions;
-        $user=Auth::user();
+        $permissions = $request->permissions;
+        $user = Auth::user();
         // dd($permissions->id);
-        foreach($permissions as $permission){
+        foreach ($permissions as $permission) {
             DB::table('model_has_permissions')->insert([
                 'permission_id' => $permission,
                 'model_type' => get_class($user), // Assuming User model
                 'model_id' => $user->id,
             ]);
-            $roles=$role->givePermissionTo($permission);
+            $roles = $role->givePermissionTo($permission);
         }
 
-      
+
         return redirect()->route('roles.index')->with('success', 'Permissions updated successfully.');
     }
-    public function destroy(Role $role)
+    public function destroy($id)
     {
-        $role->delete();
-        return redirect()->route('roles.index')->with('success', 'Role deleted successfully.');
+        $role = Role::find($id);
+        if ($role) {
+            $role->delete();
+        }
+        return response()->json(['success' => true]);
     }
 }
